requirements.txt
alembic==1.4.2
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Dance==3.0.0
Flask-Environ==0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-OAuth==0.12
Flask-OAuthlib==0.9.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
gunicorn~=20.0.4
Jinja2==2.11.2
lazy==1.4
Mako==1.1.3
MarkupSafe==1.1.1
oauth2==1.9.0.post1
oauthlib==2.1.0
Pillow==7.2.0
PyMySQL==0.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
urllib3==1.25.10
URLObject==2.4.3
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.1
